+++ seqr/utils/search/elasticsearch/es_utils.py
@@ -2,6 +2,7 @@ from collections import defaultdict
 from elasticsearch import Elasticsearch
 from elasticsearch.exceptions import ConnectionError as EsConnectionError, TransportError
 import elasticsearch_dsl
+from urllib3.connectionpool import connection_from_url
 
 from seqr.models import Sample
 from seqr.utils.redis_utils import safe_redis_get_json, safe_redis_set_json
@@ -10,7 +11,7 @@ from seqr.utils.search.elasticsearch.es_gene_agg_search import EsGeneAggSearch
 from seqr.utils.search.elasticsearch.es_search import EsSearch, get_compound_het_page
 from seqr.views.utils.json_utils import  _to_camel_case
 from settings import ELASTICSEARCH_SERVICE_HOSTNAME, ELASTICSEARCH_SERVICE_PORT, ELASTICSEARCH_CREDENTIALS, \
-    ELASTICSEARCH_PROTOCOL, ES_SSL_CONTEXT
+    ELASTICSEARCH_PROTOCOL, ES_SSL_CONTEXT, KIBANA_SERVER
 
 
 class InvalidIndexException(Exception):
@@ -65,6 +66,12 @@ def get_es_client(timeout=60, **kwargs):
 def ping_elasticsearch():
     if not get_es_client(timeout=3, max_retries=0).ping():
         raise ValueError('No response from elasticsearch ping')
+
+
+def ping_kibana():
+    resp = connection_from_url('http://{}'.format(KIBANA_SERVER)).urlopen('HEAD', '/status', timeout=3, retries=3)
+    if resp.status >= 400:
+        raise ValueError('Kibana Error {}: {}'.format(resp.status, resp.reason))
 
 
 SAMPLE_FIELDS_LIST = ['samples', 'samples_num_alt_1']
@@ -255,31 +262,35 @@ def _get_es_indices(client):
     return indices, seqr_index_projects
 
 
-def get_es_variants_for_variant_ids(samples, variant_ids, user, dataset_type=None, return_all_queried_families=False):
+def get_es_variants_for_variant_ids(samples, genome_version, variant_ids, user, dataset_type=None, return_all_queried_families=False):
     variants = EsSearch(
-        samples, user=user, return_all_queried_families=return_all_queried_families,
+        samples, genome_version, user=user, return_all_queried_families=return_all_queried_families,
     ).filter_by_variant_ids(variant_ids)
     if dataset_type:
         variants = variants.update_dataset_type(dataset_type)
     return variants.search(num_results=len(variant_ids))
 
 
-def get_es_variants(samples, search, user, previous_search_results, sort=None, page=None, num_results=None,
+def get_es_variants(samples, search, user, previous_search_results, genome_version, sort=None, page=None, num_results=None,
                     gene_agg=False, skip_genotype_filter=False):
     es_search_cls = EsGeneAggSearch if gene_agg else EsSearch
 
     es_search = es_search_cls(
         samples,
+        genome_version,
         previous_search_results=previous_search_results,
         user=user,
         sort=sort,
+        skipped_samples=search.get('skipped_samples'),
     )
 
     es_search.filter_variants(
-        inheritance=search.get('inheritance'), frequencies=search.get('freqs'), pathogenicity=search.get('pathogenicity'),
+        inheritance_mode=search.get('inheritance_mode'), inheritance_filter=search.get('inheritance_filter'),
+        frequencies=search.get('freqs'), pathogenicity=search.get('pathogenicity'),
         annotations=search.get('annotations'), annotations_secondary=search.get('annotations_secondary'),
         in_silico=search.get('in_silico'), quality_filter=search.get('qualityFilter'),
         custom_query=search.get('customQuery'), locus=search.get('locus'), skip_genotype_filter=skip_genotype_filter,
+        dataset_type=search.get('dataset_type'), secondary_dataset_type=search.get('secondary_dataset_type'),
         **search.get('parsedLocus')
 
     )