+++ seqr/utils/search/elasticsearch/es_utils_tests.py
@@ -11,7 +11,7 @@ from sys import maxsize
 from urllib3.exceptions import ReadTimeoutError
 from urllib3_mock import Responses
 
-from seqr.models import Family, Sample, VariantSearch, VariantSearchResults
+from seqr.models import Project, Family, Sample, VariantSearch, VariantSearchResults
 from seqr.utils.search.utils import get_single_variant, query_variants, \
     get_variant_query_gene_counts, get_variants_for_variant_ids, InvalidSearchException
 from seqr.utils.search.elasticsearch.es_search import _get_family_affected_status, _liftover_grch38_to_grch37
@@ -1114,7 +1114,7 @@ for fam_q in COMPOUND_HET_PATH_INHERITANCE_QUERY['bool']['should']:
     fam_quality_q = fam_q['bool']['must'][1]
     fam_quality_q['bool'] = {'should': [
         deepcopy(fam_quality_q),
-        {'terms': {'clinvar_clinical_significance': ['Pathogenic', 'Pathogenic/Likely_pathogenic']}}
+        {'regexp': {'clinvar_clinical_significance': '.*Pathogenic.*'}}
     ]}
 
 RECESSIVE_INHERITANCE_QUERY = {
@@ -1416,7 +1416,6 @@ class EsUtilsTest(TestCase):
             get_single_variant(self.families, '10-10334333-A-G')
         self.assertEqual(str(cm.exception), 'Variant 10-10334333-A-G not found')
 
-    @mock.patch('seqr.utils.search.elasticsearch.es_search.MAX_NO_LOCATION_COMP_HET_FAMILIES', 1)
     @mock.patch('seqr.utils.search.elasticsearch.es_search.MAX_COMPOUND_HET_GENES', 1)
     @mock.patch('seqr.utils.search.elasticsearch.es_gene_agg_search.MAX_COMPOUND_HET_GENES', 1)
     @mock.patch('seqr.utils.search.elasticsearch.es_search.logger')
@@ -1425,69 +1424,22 @@ class EsUtilsTest(TestCase):
         setup_responses()
         search_model = VariantSearch.objects.create(search={})
         results_model = VariantSearchResults.objects.create(variant_search=search_model)
-        results_model.families.set(Family.objects.all())
-        with self.assertRaises(InvalidSearchException) as cm:
-            query_variants(results_model)
-        self.assertEqual(
-            str(cm.exception),
-            'Searching across multiple genome builds is not supported. Remove projects with differing genome builds from search: 37 - 1kg project nåme with uniçøde, Test Reprocessed Project; 38 - Non-Analyst Project',
-        )
-
-        search_model.search = {'inheritance': {'mode': 'recessive'}, 'locus': {'rawItems': 'DDX11L1'}}
-        search_model.save()
-        results_model.families.set([family for family in self.families if family.guid == 'F000005_5'])
-        with self.assertRaises(InvalidSearchException) as cm:
-            query_variants(results_model)
-        self.assertEqual(
-            str(cm.exception), 'Inheritance based search is disabled in families with no data loaded for affected individuals',
-        )
-
-        search_model.search['inheritance']['filter'] = {'affected': {'I000007_na20870': 'N'}}
-        search_model.save()
-        results_model.families.set([family for family in self.families if family.guid == 'F000003_3'])
-        with self.assertRaises(InvalidSearchException) as cm:
-            query_variants(results_model)
-        self.assertEqual(
-            str(cm.exception),
-            'Inheritance based search is disabled in families with no data loaded for affected individuals',
-        )
-
-        search_model.search['inheritance']['filter'] = {'genotype': {'I000004_hg00731': 'ref_ref'}}
-        search_model.save()
-        with self.assertRaises(InvalidSearchException) as cm:
-            query_variants(results_model)
-        self.assertEqual(str(cm.exception), 'Invalid custom inheritance')
-
-        search_model.search['inheritance']['filter'] = {}
-        search_model.search['annotations'] = {'structural': ['DEL']}
-        search_model.save()
-        with self.assertRaises(InvalidSearchException) as cm:
-            query_variants(results_model)
-        error = 'Unable to search against dataset type "SV". This may be because inheritance based search is disabled in families with no loaded affected individuals'
-        self.assertEqual(str(cm.exception), error)
 
         results_model.families.set(self.families)
-        with self.assertRaises(InvalidSearchException) as cm:
-            query_variants(results_model, page=200)
-        self.assertEqual(str(cm.exception), 'Unable to load more than 10000 variants (20000 requested)')
-
-        search_model.search = {'inheritance': {'mode': 'compound_het'}}
-        search_model.save()
+        Sample.objects.filter(elasticsearch_index=INDEX_NAME).update(elasticsearch_index=HG38_INDEX_NAME)
         with self.assertRaises(InvalidSearchException) as cm:
             query_variants(results_model)
         self.assertEqual(
-            str(cm.exception),
-            'Annotations must be specified to search for compound heterozygous variants')
+            str(cm.exception), 'The following indices do not have the expected genome version 37: test_index_hg38 (38)',
+        )
+        Sample.objects.filter(elasticsearch_index=HG38_INDEX_NAME).update(elasticsearch_index=INDEX_NAME)
 
-        search_model.search['annotations'] = {'frameshift': ['frameshift_variant']}
-        search_model.save()
-        with self.assertRaises(InvalidSearchException) as cm:
-            query_variants(results_model)
-        self.assertEqual(
-            str(cm.exception),
-            'Location must be specified to search for compound heterozygous variants across many families')
-
-        search_model.search['locus'] = {'rawItems': 'DDX11L1'}
+        results_model.families.set(self.families)
+        search_model.search = {
+            'inheritance': {'mode': 'compound_het'},
+            'locus': {'rawItems': 'DDX11L1'},
+            'annotations': {'frameshift': ['frameshift_variant']},
+        }
         search_model.save()
         with self.assertRaises(InvalidSearchException) as cm:
             query_variants(results_model)
@@ -1546,9 +1498,8 @@ class EsUtilsTest(TestCase):
             query_variants(results_model)
         self.assertEqual(str(cm.exception), 'Could not find expected indices: test_index_sv, test_index_mito_wgs, test_index')
 
-    @mock.patch('seqr.utils.search.utils.MAX_VARIANTS')
     @urllib3_responses.activate
-    def test_get_es_variants(self, mock_max_variants):
+    def test_get_es_variants(self):
         setup_responses()
         # Testing mito indices is done in other tests, it is helpful to have a strightforward single datatype test
         Sample.objects.get(elasticsearch_index=MITO_WGS_INDEX_NAME).delete()
@@ -1563,7 +1514,7 @@ class EsUtilsTest(TestCase):
         self.assertEqual(total_results, 5)
 
         self.assertCachedResults(results_model, {'all_results': variants, 'total_results': 5})
-        self.assertTrue('index_metadata__{},{},{}'.format(INDEX_NAME, MITO_WGS_INDEX_NAME, SV_INDEX_NAME) in REDIS_CACHE)
+        self.assertTrue('index_metadata__{},{}'.format(INDEX_NAME, MITO_WGS_INDEX_NAME) in REDIS_CACHE)
 
         self.assertExecutedSearch(filters=[ANNOTATION_QUERY, ALL_INHERITANCE_QUERY])
 
@@ -1589,16 +1540,16 @@ class EsUtilsTest(TestCase):
 
         # test load_all
         setup_responses()
-        mock_max_variants.__int__.return_value = 100
-        variants, _ = query_variants(results_model, page=1, num_results=2, load_all=True)
+        with mock.patch('seqr.utils.search.utils.MAX_VARIANTS', 100):
+            variants, _ = query_variants(results_model, page=1, num_results=2, load_all=True)
         self.assertExecutedSearch(filters=[ANNOTATION_QUERY, ALL_INHERITANCE_QUERY], start_index=4, size=1)
         self.assertEqual(len(variants), 5)
         self.assertListEqual(variants, PARSED_VARIANTS + PARSED_VARIANTS + PARSED_VARIANTS[:1])
 
         # test does not re-fetch once all loaded
         urllib3_responses.reset()
-        mock_max_variants.__int__.return_value = 1
-        variants, _ = query_variants(results_model, page=1, num_results=2, load_all=True)
+        with mock.patch('seqr.utils.search.utils.MAX_VARIANTS', 1):
+            variants, _ = query_variants(results_model, page=1, num_results=2, load_all=True)
         self.assertEqual(len(variants), 5)
         self.assertListEqual(variants, PARSED_VARIANTS + PARSED_VARIANTS + PARSED_VARIANTS[:1])
 
@@ -1624,7 +1575,7 @@ class EsUtilsTest(TestCase):
                 'gnomad_genomes': {'af': 0.01, 'hh': 3},
                 'topmed': {'ac': 2, 'af': None},
             },
-            'qualityFilter': {'min_ab': 10, 'min_gq': 15, 'vcf_filter': 'pass'},
+            'qualityFilter': {'min_ab': 10, 'min_gq': 15, 'vcf_filter': 'pass', 'affected_only': True},
             'in_silico': {'cadd': '11.5', 'sift': 'D'},
             'inheritance': {'mode': 'de_novo'},
             'customQuery': {'term': {'customFlag': 'flagVal'}},
@@ -1742,10 +1693,8 @@ class EsUtilsTest(TestCase):
                                 {'range': {'gnomad_genomes_AF_POPMAX_OR_GLOBAL': {'lte': 0.05}}}
                             ]
                         }},
-                        {'terms': {
-                            'clinvar_clinical_significance': [
-                                'Likely_pathogenic', 'Pathogenic', 'Pathogenic/Likely_pathogenic',
-                            ]
+                        {'regexp': {
+                            'clinvar_clinical_significance': '.*Likely_pathogenic.*|.*Pathogenic.*',
                         }}
                     ]
                 }},
@@ -1769,11 +1718,8 @@ class EsUtilsTest(TestCase):
                                 'intergenic_variant',
                             ]
                         }},
-                        {'terms': {
-                            'clinvar_clinical_significance': [
-                                'Conflicting_interpretations_of_pathogenicity', 'Likely_pathogenic', 'Pathogenic',
-                                'Pathogenic/Likely_pathogenic', 'Uncertain_significance', 'not_provided', 'other',
-                            ]
+                        {'regexp': {
+                            'clinvar_clinical_significance': '.*Likely_pathogenic.*|.*Pathogenic.*|Conflicting_interpretations_of_pathogenicity.*|~((.*[Bb]enign.*)|(.*[Pp]athogenic.*))',
                         }},
                         {'terms': {'hgmd_class': ['DM', 'DM?']}},
                         {'range': {'splice_ai_delta_score': {'gte': 0.8}}},
@@ -1816,41 +1762,9 @@ class EsUtilsTest(TestCase):
                                     {'term': {'samples_gq_0_to_5': 'HG00731'}},
                                     {'term': {'samples_gq_5_to_10': 'HG00731'}},
                                     {'term': {'samples_gq_10_to_15': 'HG00731'}},
-                                    {'term': {'samples_gq_0_to_5': 'HG00732'}},
-                                    {'term': {'samples_gq_5_to_10': 'HG00732'}},
-                                    {'term': {'samples_gq_10_to_15': 'HG00732'}},
-                                    {'term': {'samples_gq_0_to_5': 'HG00733'}},
-                                    {'term': {'samples_gq_5_to_10': 'HG00733'}},
-                                    {'term': {'samples_gq_10_to_15': 'HG00733'}},
                                 ],
-                                'must': [
-                                    {'bool': {
-                                        'minimum_should_match': 1,
-                                        'should': [
-                                            {'bool': {
-                                                'must_not': [
-                                                    {'term': {'samples_ab_0_to_5': 'HG00732'}},
-                                                    {'term': {'samples_ab_5_to_10': 'HG00732'}},
-                                                ]
-                                            }},
-                                            {'bool': {'must_not': [{'term': {'samples_num_alt_1': 'HG00732'}}]}}
-                                        ]
-                                    }},
-                                    {'bool': {
-                                        'minimum_should_match': 1,
-                                        'should': [
-                                            {'bool': {
-                                                'must_not': [
-                                                    {'term': {'samples_ab_0_to_5': 'HG00733'}},
-                                                    {'term': {'samples_ab_5_to_10': 'HG00733'}},
-                                                ]
-                                            }},
-                                            {'bool': {'must_not': [{'term': {'samples_num_alt_1': 'HG00733'}}]}}
-                                        ]
-                                    }},
-                                ]
-                            }}, {'terms': {
-                                'clinvar_clinical_significance': ['Likely_pathogenic', 'Pathogenic', 'Pathogenic/Likely_pathogenic']
+                            }}, {'regexp': {
+                                'clinvar_clinical_significance': '.*Likely_pathogenic.*|.*Pathogenic.*'
                             }}]}}
                         ],
                     }},
@@ -1876,8 +1790,8 @@ class EsUtilsTest(TestCase):
                                     {'term': {'samples_gq_5_to_10': 'NA20870'}},
                                     {'term': {'samples_gq_10_to_15': 'NA20870'}},
                                 ]
-                            }}, {'terms': {
-                                'clinvar_clinical_significance': ['Likely_pathogenic', 'Pathogenic', 'Pathogenic/Likely_pathogenic']
+                            }}, {'regexp': {
+                                'clinvar_clinical_significance': '.*Likely_pathogenic.*|.*Pathogenic.*',
                             }}]}}
                         ],
                         '_name': 'F000003_3'
@@ -1979,10 +1893,8 @@ class EsUtilsTest(TestCase):
 
         variants, _ = query_variants(results_model, num_results=5)
         self.assertListEqual(variants, [PARSED_SV_VARIANT] + PARSED_NO_CONSEQUENCE_FILTER_VARIANTS + [PARSED_MITO_VARIANT])
-        path_filter = {'terms': {
-            'clinvar_clinical_significance': [
-                'Pathogenic', 'Pathogenic/Likely_pathogenic'
-            ]
+        path_filter = {'regexp': {
+            'clinvar_clinical_significance':  '.*Pathogenic.*'
         }}
         self.assertExecutedSearches([
             dict(filters=[path_filter], start_index=0, size=5, index=SV_INDEX_NAME),
@@ -2099,7 +2011,7 @@ class EsUtilsTest(TestCase):
         })
 
         annotation_query = {'bool': {'should': [
-            {'terms': {'clinvar_clinical_significance': ['Pathogenic', 'Pathogenic/Likely_pathogenic']}},
+            {'regexp': {'clinvar_clinical_significance': '.*Pathogenic.*'}},
             {'terms': {'hgmd_class': ['DM']}},
             {'range': {'splice_ai_delta_score': {'gte': 0.5}}},
             {'terms': {'transcriptConsequenceTerms': ['frameshift_variant', 'intron']}},
@@ -2387,8 +2299,67 @@ class EsUtilsTest(TestCase):
                             ]
                         }},
                     ]
-                    }},
-                 ],
+                }},
+                         ],
+                gene_aggs=True,
+                start_index=0,
+                size=1,
+                index=','.join([INDEX_NAME, SV_INDEX_NAME]),
+            ),
+            dict(
+                filters=[
+                    annotation_secondary_query,
+                    {'bool': {'_name': 'F000003_3', 'must': [{'term': {'samples_num_alt_1': 'NA20870'}}]}},
+                ],
+                gene_aggs=True,
+                start_index=0,
+                size=1
+            ),
+        ])
+
+    @urllib3_responses.activate
+    def test_multi_datatype_secondary_annotations_comp_het_get_es_variants(self):
+        setup_responses()
+        search_model = VariantSearch.objects.create(search={
+            'annotations': {'structural': ['DEL'], 'SCREEN': ['dELS']},
+            'annotations_secondary': {'structural_consequence': ['LOF']},
+            'inheritance': {'mode': 'compound_het'},
+        })
+        results_model = VariantSearchResults.objects.create(variant_search=search_model)
+        results_model.families.set(self.families)
+
+        query_variants(results_model, num_results=10)
+
+        annotation_secondary_query = {'bool': {'should': [
+            {'terms': {'transcriptConsequenceTerms': ['DEL', 'LOF']}},
+            {'terms': {'screen_region_type': ['dELS']}}]}}
+
+        self.assertExecutedSearches([
+            dict(
+                filters=[annotation_secondary_query, {'bool': {
+                    '_name': 'F000002_2',
+                    'must': [
+                        {'bool': {
+                            'should': [
+                                {'bool': {
+                                    'minimum_should_match': 1,
+                                    'must_not': [
+                                        {'term': {'samples_no_call': 'HG00732'}},
+                                        {'term': {'samples_num_alt_2': 'HG00732'}},
+                                        {'term': {'samples_no_call': 'HG00733'}},
+                                        {'term': {'samples_num_alt_2': 'HG00733'}}
+                                    ],
+                                    'should': [
+                                        {'term': {'samples_num_alt_1': 'HG00731'}},
+                                        {'term': {'samples_num_alt_2': 'HG00731'}},
+                                    ]
+                                }},
+                                {'term': {'samples': 'HG00731'}},
+                            ]
+                        }},
+                    ]
+                }},
+                         ],
                 gene_aggs=True,
                 start_index=0,
                 size=1,
@@ -2780,6 +2751,7 @@ class EsUtilsTest(TestCase):
 
         _, total_results = query_variants(results_model, num_results=2)
         self.assertEqual(total_results, 14)
+        self.assertTrue('index_metadata__{},{},{}'.format(INDEX_NAME, MITO_WGS_INDEX_NAME, SV_INDEX_NAME) in REDIS_CACHE)
 
         gene_filter = {'terms': {'geneIds': ['ENSG00000228198']}}
         prefilter_search = dict(
@@ -2809,8 +2781,9 @@ class EsUtilsTest(TestCase):
                     '_name': 'F000002_2'
                 }}
             ], start_index=0, size=2, index=SV_INDEX_NAME)
-        self.assertEqual(len(urllib3_responses.calls), 2)
-        self.assertExecutedSearch(call_index=0, **prefilter_search)
+        num_calls = 3
+        self.assertEqual(len(urllib3_responses.calls), num_calls)
+        self.assertExecutedSearch(call_index=num_calls-2, **prefilter_search)
         # Search total is greater than returned hits, so proceed with regular multi-search
         self.assertExecutedSearches([
             sv_search,
@@ -2884,8 +2857,9 @@ class EsUtilsTest(TestCase):
 
         _, total_results = query_variants(results_model, num_results=2)
         self.assertEqual(total_results, 1)
-        self.assertEqual(len(urllib3_responses.calls), 4)
-        self.assertExecutedSearch(call_index=2, **prefilter_search)
+        num_calls += 2
+        self.assertEqual(len(urllib3_responses.calls), num_calls)
+        self.assertExecutedSearch(call_index=num_calls-2, **prefilter_search)
         sv_search['query'] = [{'ids': {'values': ['prefix_19107_DEL']}}]
         self.assertExecutedSearches([sv_search])
 
@@ -2898,7 +2872,7 @@ class EsUtilsTest(TestCase):
         _, total_results = query_variants(results_model, num_results=2)
         self.assertEqual(total_results, 0)
         # Only the prefliter search is run, no multi-search
-        self.assertEqual(len(urllib3_responses.calls), 5)
+        self.assertEqual(len(urllib3_responses.calls), num_calls + 1)
         self.assertExecutedSearch(**prefilter_search)
 
 
@@ -2967,6 +2941,7 @@ class EsUtilsTest(TestCase):
         results_model = VariantSearchResults.objects.create(variant_search=search_model)
         results_model.families.set(Family.objects.filter(guid__in=['F000011_11']))
 
+        Project.objects.filter(guid='R0003_test').update(genome_version='38')
         Sample.objects.filter(elasticsearch_index=SECOND_INDEX_NAME).update(elasticsearch_index=HG38_INDEX_NAME)
 
         mock_liftover.side_effect = Exception()
@@ -3322,10 +3297,6 @@ class EsUtilsTest(TestCase):
                 }
             }, 'xpos', 'variantId'])
 
-        with self.assertRaises(InvalidSearchException) as se:
-            query_variants(results_model, sort='prioritized_gene', num_results=2)
-        self.assertEqual(str(se.exception), 'Phenotype sort is only supported for single-family search.')
-
         results_model.families.set(Family.objects.filter(guid='F000001_1'))
         query_variants(results_model, sort='prioritized_gene', num_results=2)
         family_sample_filter = {'bool': {'_name': 'F000001_1', 'must': mock.ANY}}
@@ -3355,7 +3326,7 @@ class EsUtilsTest(TestCase):
 
         def _execute_inheritance_search(
                 mode=None, inheritance_filter=None, expected_filter=None, expected_comp_het_filter=None,
-                dataset_type=Sample.DATASET_TYPE_VARIANT_CALLS, **kwargs):
+                quality_filter=None, expected_quality_filter=None, dataset_type=Sample.DATASET_TYPE_VARIANT_CALLS, **kwargs):
             _set_cache(cache_key, None)
             annotations = {'frameshift': ['frameshift_variant']} if dataset_type == Sample.DATASET_TYPE_VARIANT_CALLS \
                 else {'structural': ['DEL', 'gCNV_DEL']}
@@ -3363,6 +3334,8 @@ class EsUtilsTest(TestCase):
                 'inheritance': {'mode': mode, 'filter': inheritance_filter},
                 'annotations': annotations,
             }
+            if quality_filter:
+                search_model.search['qualityFilter'] = quality_filter
             search_model.save()
             query_variants(results_model, num_results=2)
 
@@ -3377,8 +3350,11 @@ class EsUtilsTest(TestCase):
                         annotation_query,  {'bool': {'_name': 'F000002_2', 'must': [expected_filter]}}])
                 ])
             else:
+                filters = [expected_filter]
+                if expected_quality_filter:
+                    filters.append(expected_quality_filter)
                 self.assertExecutedSearch(index=index, filters=[
-                    annotation_query, {'bool': {'_name': 'F000002_2', 'must': [expected_filter]}}], **kwargs)
+                    annotation_query, {'bool': {'_name': 'F000002_2', 'must': filters}}], **kwargs)
 
         # custom genotype
         _execute_inheritance_search(
@@ -3438,6 +3414,29 @@ class EsUtilsTest(TestCase):
                 ]
             }
         })
+
+        _execute_inheritance_search(mode='de_novo', inheritance_filter={'affected': custom_affected}, expected_filter={
+            'bool': {
+                'minimum_should_match': 1,
+                'must_not': [
+                    {'term': {'samples_no_call': 'HG00731'}},
+                    {'term': {'samples_num_alt_1': 'HG00731'}},
+                    {'term': {'samples_num_alt_2': 'HG00731'}},
+                    {'term': {'samples_no_call': 'HG00733'}},
+                    {'term': {'samples_num_alt_1': 'HG00733'}},
+                    {'term': {'samples_num_alt_2': 'HG00733'}}
+                ],
+                'should': [
+                    {'term': {'samples_num_alt_1': 'HG00732'}},
+                    {'term': {'samples_num_alt_2': 'HG00732'}}
+                ]
+            }
+        }, quality_filter={'affected_only': True, 'min_gq': 10}, expected_quality_filter={
+            'bool': {'must_not': [
+                {'term': {'samples_gq_0_to_5': 'HG00732'}},
+                {'term': {'samples_gq_5_to_10': 'HG00732'}},
+            ]}}
+        )
 
         _execute_inheritance_search(
             mode='de_novo', inheritance_filter={'affected': custom_multi_affected}, expected_filter={'bool': {
@@ -3563,7 +3562,7 @@ class EsUtilsTest(TestCase):
                     {'term': {'samples_num_alt_2': 'HG00733'}}
                 ],
                 'must': [
-                    {'match': {'contig': 'X'}},
+                    {'range': {'xpos': {'gte': 23000000001, 'lte': 24000000001}}},
                     {'term': {'samples_num_alt_2': 'HG00731'}},
                 ]
             }
@@ -3582,7 +3581,7 @@ class EsUtilsTest(TestCase):
                     {'term': {'samples_cn_0': 'HG00733'}},
                     {'term': {'samples_cn_gte_4': 'HG00733'}},
                 ],
-                'must': [{'match': {'contig': 'X'}}],
+                'must': [{'range': {'xpos': {'gte': 23000000001, 'lte': 24000000001}}},],
             }
         }
 
@@ -3595,7 +3594,7 @@ class EsUtilsTest(TestCase):
                     {'term': {'samples_num_alt_2': 'HG00733'}}
                 ],
                 'must': [
-                    {'match': {'contig': 'X'}},
+                    {'range': {'xpos': {'gte': 23000000001, 'lte': 24000000001}}},
                     {'term': {'samples_num_alt_2': 'HG00732'}},
                 ]
             }
@@ -3639,8 +3638,3 @@ class EsUtilsTest(TestCase):
                 ]
             }
         })
-
-        # Affected specified with no other inheritance
-        with self.assertRaises(InvalidSearchException) as cm:
-            _execute_inheritance_search(inheritance_filter={'affected': custom_affected})
-        self.assertEqual(str(cm.exception), 'Inheritance must be specified if custom affected status is set')